modified to be more verbose and return proper error code upon exit

+++ buildroot-external/patches/occu/create_patches.sh
@@ -4,6 +4,8 @@
 # patchset files by generating diff output of orig<>real file differences in
 # all the 0000-XXXX folders
 
+set -e
+
 export LC_ALL=C
 
 patchdirs=$(find . -maxdepth 1 -type d -regex "\\./[0-9][0-9][0-9][0-9]-.*" | sort)
@@ -13,10 +15,14 @@ rm -f ./*.patch
 for dir in ${patchdirs}; do
   (
     cd "${dir}" || exit 1
+    echo $(basename ${dir})
     origfiles=$(find occu -name "*.orig" -type f -print | sort)
+    if [ -z "${origfiles}" ]; then
+      exit 1
+    fi
     rm -f "../${dir}.patch"
     for file in ${origfiles}; do
-      diff -u --label="${file}" --label="${file%.orig}" "${file}" "${file%.orig}" >>"../${dir}.patch"
+      diff -u --label="${file}" --label="${file%.orig}" "${file}" "${file%.orig}" >>"../${dir}.patch" || true
     done
   )
 done
